refactor: move update_check out of operations into dist
CLI self-update checking is not a library operation. Fold
`operations/update_check.rs` into `dist.rs`, which already owns the
`dist_supports_update_check`/`dist_check_for_updates` helpers it calls,
and update the single call site in `main.rs`.
