+++ crates/icp-cli/src/operations/mod.rs
@@ -14,5 +14,4 @@ pub(crate) mod sync;
 pub(crate) mod token;
 
 pub(crate) mod misc;
-pub(crate) mod update_check;
 pub(crate) mod wasm;
